Create invisible.py
The "Cloak of Invisibility" project is inspired by the concept of making objects or individuals seem invisible to the human eye, a concept often depicted within science fiction and fantasy. In computer science terms, this project often makes reference to the use of techniques that relate to image processing and computer vision to give an illusion of invisibility.